fix: resolve all clippy warnings, remove dead code, add workspace lints
- Fix all 36 clippy warnings (auto-fix + manual fixes)
- Remove all 8 #[allow(dead_code)] annotations and delete underlying dead code
- Add workspace-level lint rules in Cargo.toml (deny: dead_code, unused_imports, unreachable_code)
- Remove dead fields/methods from App, RouterEngine, RequestParser, ClipboardHandler

+++ src/ui/history/dialog.rs
@@ -91,7 +91,7 @@ fn chrono_local_offset_secs() -> i64 {
             let now = libc::time(std::ptr::null_mut());
             let mut tm = MaybeUninit::<libc::tm>::uninit();
             libc::localtime_r(&now, tm.as_mut_ptr());
-            (*tm.as_ptr()).tm_gmtoff as i64
+            (*tm.as_ptr()).tm_gmtoff
         }
     }
     #[cfg(not(unix))]
